+++ packages/api/src/augment/consts.ts
@@ -859,6 +859,7 @@ declare module '@polkadot/api/types/consts' {
       [key: string]: Codec;
     };
     vesting: {
+      maxVestingSchedules: u32 & AugmentedConst<ApiType>;
       /**
        * The minimum amount transferred to call `vested_transfer`.
        **/