+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## master
+
+Changes:
+
+- Bump static metadata for latest Substrate, Polkadot & Kusama
+
+
 ## 5.7.1 Aug 29, 2021
 
 Upgrade priority: Low.